Stats: averageResponse type changed to long and now represents time in milliseconds

+++ src/main/kotlin/ru/redenergy/report/common/Stats.kt
@@ -5,6 +5,6 @@ package ru.redenergy.report.common
  *
  * @param tickets: Contains amount of tickets in each category (TicketReason: amount)
  * @param activeUsers: Contains five of most active users (UserName: amount)
- * @param averageTime: Average response time for a ticket (int minutes)
+ * @param averageTime: Average response time for a ticket (in milliseconds), if 0
  */
-data class Stats(val tickets: Map<TicketReason, Int>, val activeUsers: Map<String, Int>, val averageTime: Int)
+data class Stats(val tickets: Map<TicketReason, Int>, val activeUsers: Map<String, Int>, val averageTime: Long)
